scaled play animation routes and added play viz II

+++ playbooks/playbook.py
@@ -32,8 +32,8 @@ class PlaybookType(Enum):
     ALTERNATE = 3
 
 class Point(NamedTuple):
-    x: int  # normalized
-    y: int  # normalized
+    x: float
+    y: float
 
 @dataclass
 class Route:
